feat: custom types
This introduces the concept of custom types. Our services often need
to set multiple values as env vars, while they often are represented as
one value in the infrastructure.

Take the Pub/Sub topic for example. If publishing a message to a topic
in a different project, often the project id and topic id is passed as
two environment variables. This addition will allow us to pass the
single value format, `TOPIC_ID=projects/project-id/topic/topic-id`, and
nicely decode it into a struct containing both values.

Co-authored-by: Henrik Hedlund <henrik.hedlund@remarkable.no>