feat!: Treat empty strings as missing
This changes so that there is not difference between missing
environment variables, and variable set to an empty string.
The consequences here are that required fields always must
contains a value, and hat default values always will override
empty environment variables.

More subtle consequences includes that types such as maps will
behave the same way if an environment variable is set to empty
string, as if it is missing, i.e. not be initiated in the first
place.

This is a breaking change.